requirements.txt
flask==2.3.3
flask-sqlalchemy==3.1.1
pytest==7.4.0
pytest-mock==3.11.1
python-dotenv==1.0.0